projeto-final-pi/requirements.txt
Flask==3.0.3
mysql_connector_repackaged==0.3.1
pip==24.0
